+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.2.6 2026
 
+* [Add Protyle fullscreen state APIs](https://github.com/siyuan-note/siyuan/issues/18909)
 * [Support multiple editor font family configuration types](https://github.com/siyuan-note/siyuan/issues/16923)
 * [Add frontend asset upload plugin event types](https://github.com/siyuan-note/siyuan/issues/18878)
 * [Add editor breadcrumb button plugin APIs](https://github.com/siyuan-note/siyuan/issues/18856)